add pereg command 37,38

+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.app/src/main/java/nts/uk/ctx/at/record/app/command/remainingnumber/annleagrtremnum/AddAnnLeaGrantRemnNumPeregCommand.java
@@ -1,5 +1,7 @@
 package nts.uk.ctx.at.record.app.command.remainingnumber.annleagrtremnum;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 import nts.arc.time.GeneralDate;
 import nts.uk.shr.pereg.app.PeregEmployeeId;
@@ -28,42 +30,42 @@ public class AddAnnLeaGrantRemnNumPeregCommand {
 	 * 年休期限切れ状態
 	 */
 	@PeregItem("IS00387")
-	private int expirationStatus;
+	private BigDecimal expirationStatus;
 
 	/**
 	 * 付与日数
 	 */
 	@PeregItem("IS00390")
-	private Double grantDays;
+	private BigDecimal grantDays;
 
 	/**
 	 * 付与時間
 	 */
 	@PeregItem("IS00391")
-	private Integer grantMinutes;
+	private BigDecimal grantMinutes;
 
 	/**
 	 * 使用日数
 	 */
 	@PeregItem("IS00393")
-	private Double usedDays;
+	private BigDecimal usedDays;
 
 	/**
 	 * 使用時間
 	 */
 	@PeregItem("IS00394")
-	private Integer usedMinutes;
+	private BigDecimal usedMinutes;
 
 	/**
 	 * 残日数
 	 */
 	@PeregItem("IS00396")
-	private Double remainingDays;
+	private BigDecimal remainingDays;
 
 	/**
 	 * 残時間
 	 */
 	@PeregItem("IS00397")
-	private Integer remainingMinutes;
+	private BigDecimal remainingMinutes;
 
 }
